# apps/h5/openapi.json
{
  "components": {
    "schemas": {
...
      "CaptureAcceptedResponse": {
        "properties": {
          "capture_id": {
            "format": "uuid",
            "title": "Capture Id",
            "type": "string"
          },
          "events_url": {
            "title": "Events Url",
            "type": "string"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "look_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Look Id"
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "status_url": {
            "title": "Status Url",
            "type": "string"
          }
        },
        "required": [
          "capture_id",
          "job_id",
          "state",
          "status_url",
          "events_url"
        ],
        "title": "CaptureAcceptedResponse",
        "type": "object"
      },
      "CaptureSourceKind": {
        "enum": [
          "upload",
          "camera",
          "feed"
        ],
        "title": "CaptureSourceKind",
        "type": "string"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "FeedCaptureIntent": {
        "enum": [
          "item_selections",
          "whole_outfit"
        ],
        "title": "FeedCaptureIntent",
        "type": "string"
      },
      "FeedFrameContextBody": {
        "properties": {
          "frame_height": {
            "exclusiveMinimum": 0.0,
            "title": "Frame Height",
            "type": "integer"
          },
          "frame_width": {
            "exclusiveMinimum": 0.0,
            "title": "Frame Width",
            "type": "integer"
          },
          "intent": {
            "$ref": "#/components/schemas/FeedCaptureIntent",
            "default": "item_selections"
          },
          "selections": {
            "items": {
              "$ref": "#/components/schemas/FeedSelectionBody"
            },
            "maxItems": 8,
            "minItems": 1,
            "title": "Selections",
            "type": "array"
          },
          "timestamp_ms": {
            "minimum": 0.0,
            "title": "Timestamp Ms",
            "type": "integer"
          },
          "video_ref": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Video Ref",
            "type": "string"
          }
        },
        "required": [
          "video_ref",
          "timestamp_ms",
          "frame_width",
          "frame_height",
          "selections"
        ],
        "title": "FeedFrameContextBody",
        "type": "object"
      },
      "FeedSelectionBody": {
        "properties": {
          "polygon": {
            "items": {
              "$ref": "#/components/schemas/NormalizedPointBody"
            },
            "maxItems": 256,
            "minItems": 3,
            "title": "Polygon",
            "type": "array"
          },
          "selection_key": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Selection Key",
            "type": "string"
          }
        },
        "required": [
          "selection_key",
          "polygon"
        ],
        "title": "FeedSelectionBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobState": {
        "enum": [
          "queued",
          "processing",
          "partial",
          "ready",
          "error"
        ],
        "title": "JobState",
        "type": "string"
      },
...
      "NormalizedPointBody": {
        "properties": {
          "x": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "X",
            "type": "number"
          },
          "y": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Y",
            "type": "number"
          }
        },
        "required": [
          "x",
          "y"
        ],
        "title": "NormalizedPointBody",
        "type": "object"
      },
      "OwnershipState": {
        "enum": [
          "owned",
          "inspiration"
        ],
        "title": "OwnershipState",
        "type": "string"
      },
...
      "SubmitCaptureBody": {
        "properties": {
          "feed_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FeedFrameContextBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "object_key": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Object Key",
            "type": "string"
          },
          "ownership": {
            "$ref": "#/components/schemas/OwnershipState"
          },
          "sha256": {
            "maxLength": 64,
            "minLength": 64,
            "title": "Sha256",
            "type": "string"
          },
          "source_kind": {
            "$ref": "#/components/schemas/CaptureSourceKind"
          }
        },
        "required": [
          "object_key",
          "sha256",
          "source_kind",
          "ownership"
        ],
        "title": "SubmitCaptureBody",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Versioned capture and digital wardrobe capabilities.",
    "title": "StyleCapture Product API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/captures": {
      "post": {
        "operationId": "submit_capture_v1_captures_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "stylecapture_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stylecapture Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitCaptureBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request cannot be processed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A valid product session is required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested resource was not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request conflicts with stored state"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The upload token has expired"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The upload exceeds the allowed size"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The media type is not supported"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request violates the API contract"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Processing is temporarily unavailable"
          }
        },
        "summary": "Submit Capture"
      }
    },
...
    "/v1/session": {
      "post": {
        "operationId": "create_session_v1_session_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "stylecapture_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stylecapture Session"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "title": "Response Create Session V1 Session Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Session"
      }
    },
...
  }
}